[fixed]MeshEngine初始化出错 [fixed]BaseShader释放不完全 [TODO]glCompileShader很慢哦

+++ jni/effect/imageeffect.cpp
@@ -55,7 +55,7 @@ bool ImageEffect::loadFromRes( const char* fileName )
 
 void ImageEffect::process( Texture* src, Texture* dst )
 {
-    GLuint w,h;
+    GLfloat w,h;
     w = src->getWidth();
     h = src->getHeight();
     dst->setSize(w, h);
